Moved telemetry.go to engine package and removed test from model_test
