+++ server/engine/telemetry.go
@@ -1,8 +1,9 @@
-package model
+package engine
 
 import (
 	"fmt"
 	"server/config"
+	"server/model"
 
 	"github.com/segmentio/analytics-go/v3"
 	log "github.com/sirupsen/logrus"
@@ -11,12 +12,12 @@ import (
 
 func GetEvent(propertiesMap analytics.Properties) string {
 
-	switch status := propertiesMap[string(DeployStatus)]; status {
+	switch status := propertiesMap[string(model.DeployStatus)]; status {
 	case "succeeded":
 		return "aap.azure.installer-deploy-success"
 	case "failed":
 		return "aap.azure.installer-deploy-failed"
-	case "cancelled":
+	case "canceled":
 		return "aap.azure.installer-deploy-cancel"
 	}
 	return ""
@@ -25,7 +26,7 @@ func GetEvent(propertiesMap analytics.Properties) string {
 func BuildSegmentPropertiesMap(db *gorm.DB) analytics.Properties {
 
 	var propertiesMap = analytics.Properties{}
-	var metricData []Telemetry
+	var metricData []model.Telemetry
 	db.Find(&metricData)
 	for _, data := range metricData {
 		propertiesMap[string(data.MetricName)] = fmt.Sprintf("%v", data.MetricValue)
@@ -40,12 +41,12 @@ func PublishToSegment(db *gorm.DB) {
 		log.Errorf("Segment Write Key is missing : Not sending telemetry to Segment")
 	}
 	// set metrics in DB that are not set yet
-	SetMetric(db, ApplicationId, config.GetEnvironment().APPLICATION_ID)
+	model.SetMetric(db, model.ApplicationId, config.GetEnvironment().APPLICATION_ID)
 	//gather all metrics in a property map
 	propertiesMap := BuildSegmentPropertiesMap(db)
 	eventName := GetEvent(propertiesMap)
 	if eventName == "" {
-		log.Errorf("Unexpected value for deploy status: [%v]. Not sending telemetry to Segment.", propertiesMap[string(DeployStatus)])
+		log.Errorf("Unexpected value for deploy status: [%v]. Not sending telemetry to Segment.", propertiesMap[string(model.DeployStatus)])
 		return
 	}
